move flask related code to different files to ease imports
