+++ rl/algorithms/Algorithm.py
@@ -1,0 +1,7 @@
+from rl.api import logger
+
+class Algorithm:
+    def __init__(self) -> None:
+        pass
+    
+
