+++ rl/api/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask
+
+from rl.logger.Logger import Logger
+
+app = Flask(__name__)
+logger = Logger()
